put mock interest lists in DummyInterestDatabase
