+++ app/src/main/java/com/github/polybooks/core/database/implementation/DummyInterestDatabase.kt
@@ -2,12 +2,41 @@ package com.github.polybooks.core.database.implementation
 
 import com.github.polybooks.core.*
 import com.github.polybooks.core.database.interfaces.InterestDatabase
-import com.github.polybooks.utils.mockCourses
-import com.github.polybooks.utils.mockFields
-import com.github.polybooks.utils.mockSemesters
 import java.util.concurrent.CompletableFuture
 
 class DummyInterestDatabase : InterestDatabase {
+
+    companion object {
+        val mockFields : List<Field> = listOf(
+            Field("Biology"),
+            Field("Computer Science"),
+            Field("Architecture"))
+
+        val mockCourses : List<Course> = listOf(
+            Course("COM-101"),
+            Course("CS-306"),
+            Course("CS-323"),
+            Course("EE-280"),
+            Course("MSE-210"),
+            Course("HUM-201"),
+            Course("DH-405"),
+            Course("ENV-444"),
+            Course("MICRO-511")
+        )
+
+        val mockSemesters : List<Semester> = listOf(
+
+            Semester("IN", "BA1"),
+            Semester("SV", "BA1"),
+            Semester("GC", "MA2"),
+            Semester("SC", "BA6"),
+            Semester("MT", "BA2"),
+            Semester("MX", "BA3"),
+            Semester("AR", "MA1"),
+            Semester("CD", "BA4"),
+            Semester("ENV", "BA5")
+        )
+    }
 
     override fun listAllFields(): CompletableFuture<List<Field>> {
         return CompletableFuture.supplyAsync {
